Move _clean_readme_summary into the pure workflow_inputs module
